Add .gitignore and update WebAPI project settings
Added a .gitignore file to exclude files generated by Visual Studio, macOS, and Windows. This is necessary to prevent unneeded files from being committed to the repository.

Also, updated the WebAPI project's launchSettings.json and Program.cs. Removed IIS settings from launchSettings.json and set the applicationUrl to run on port 10001 instead of 5125. This was done to avoid conflicts with other local applications using the same port.

In the Program.cs, added rate limiting to the API endpoint. This helps to control the rate of incoming requests and provides protection against denial-of-service attacks. Also, commented out the app.UseHttpsRedirection() line.